Add namespace to Fifo class
